# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AI饮食识别系统 API",
    "version": "1.0.0",
    "description": "基于 Spring Boot 的识别/会话/日志/用户/分析接口"
  },
  "schemes": ["http"],
  "basePath": "/",
  "consumes": ["application/json", "multipart/form-data"],
  "produces": ["application/json"],
  "paths": {
...
    "/user/user/profile/{userId}": {
      "get": {
        "summary": "获取用户信息",
        "parameters": [{ "name": "userId", "in": "path", "type": "integer", "required": true }],
        "responses": { "200": { "description": "OK", "schema": { "$ref": "#/definitions/UserProfile" } } }
      }
    },
    "/user/user/profile": {
      "put": {
        "summary": "更新用户信息",
        "parameters": [{ "in": "body", "name": "body", "required": true, "schema": { "$ref": "#/definitions/UserProfile" } }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/user/user/body/{userId}": {
      "get": {
        "summary": "获取身体信息",
        "parameters": [{ "name": "userId", "in": "path", "type": "integer", "required": true }],
        "responses": { "200": { "description": "OK", "schema": { "$ref": "#/definitions/UserBodyProfile" } } }
      }
    },
    "/user/user/body": {
      "put": {
        "summary": "更新身体信息",
        "parameters": [{ "in": "body", "name": "body", "required": true, "schema": { "$ref": "#/definitions/UserBodyProfile" } }],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
  },
  "definitions": {
...
    "UserProfile": {
      "type": "object",
      "properties": {
        "id": { "type": "integer", "format": "int64" },
        "userId": { "type": "integer", "format": "int64" },
        "nickname": { "type": "string" },
        "avatar": { "type": "string" },
        "gender": { "type": "integer", "format": "int32" }
      }
    },
    "UserBodyProfile": {
      "type": "object",
      "properties": {
        "id": { "type": "integer", "format": "int64" },
        "userId": { "type": "integer", "format": "int64" },
        "height": { "type": "number" },
        "weight": { "type": "number" },
        "bodyFat": { "type": "number" },
        "age": { "type": "integer", "format": "int32" },
        "bmi": { "type": "number" },
        "bmr": { "type": "number" },
        "targetCalories": { "type": "number" }
      }
    },
...
  }
}